Have AsyncLazy always cache values
